pyproject.toml
[project]
name = "hardware-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "fastapi>=0.115.8",
    "isort>=6.0.0",
    "openai>=1.64.0",
    "pyaudio>=0.2.14",
    "pydantic-settings>=2.8.0",
    "pydub>=0.25.1",
    "requests>=2.32.3",
    "speechrecognition>=3.14.1",
]
